requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
psycopg[binary]==3.2.13
python-docx==1.1.0
google-generativeai==0.3.2
python-pptx==0.6.23
python-dotenv==1.0.0

